requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
Werkzeug==3.0.1
pandas==2.1.4
openpyxl==3.1.2
numpy==1.26.3
